More robust checking for C11 threads
Ensure C11 threads is searched for properly in Windows. Windows does not define the `__STDC_NO_THREADS__` macro despite not having the `threads.h` header (as of Visual Studio 2022). As such, when no threads.h header is found, ensure the no_threads macro is set so the tinycthread fallback is properly reached.